pyproject.toml
[project]
name = "alden-finder"
version = "0.1.0"
description = "Aggregated in-stock Alden footwear search across authorized retailers worldwide."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Alden Finder Contributors" }]
dependencies = [
  "streamlit>=1.37",
  "httpx>=0.27",
  "selectolax>=0.3.21",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "python-dateutil>=2.9",
  "supabase>=2.5",
  "tenacity>=8.3",
  "rapidfuzz>=3.9",
  "babel>=2.15",
  "python-slugify>=8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "respx>=0.21",
]

[project.scripts]
alden-scrape = "alden_finder.scraping.runner:cli"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["alden_finder*"]

[tool.setuptools.package-data]
"alden_finder" = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
